Move db connection utils to dbconn
