+++ internal/db/dbutil/dbutil.go
@@ -8,27 +8,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/url"
-	"os"
-	"strconv"
 	"strings"
 	"time"
 
-	// Register driver
-
-	"github.com/jackc/pgconn"
-	_ "github.com/lib/pq"
-
-	"github.com/golang-migrate/migrate/v4"
-	"github.com/golang-migrate/migrate/v4/database/postgres"
-	bindata "github.com/golang-migrate/migrate/v4/source/go_bindata"
 	"github.com/hashicorp/go-multierror"
-	"github.com/inconshreveable/log15"
+	"github.com/jackc/pgconn"
 	"github.com/opentracing/opentracing-go/ext"
 	"github.com/pkg/errors"
 
 	"github.com/sourcegraph/sourcegraph/internal/trace/ot"
-	codeintelMigrations "github.com/sourcegraph/sourcegraph/migrations/codeintel"
-	frontendMigrations "github.com/sourcegraph/sourcegraph/migrations/frontend"
 )
 
 // Transaction calls f within a transaction, rolling back if any error is
@@ -77,158 +65,6 @@ type Tx interface {
 // A TxBeginner captures BeginTx method of a sql.DB
 type TxBeginner interface {
 	BeginTx(context.Context, *sql.TxOptions) (*sql.Tx, error)
-}
-
-// NewDB returns a new *sql.DB from the given dsn (data source name).
-func NewDB(dsn, app string) (*sql.DB, error) {
-	cfg, err := url.Parse(dsn)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to parse dsn")
-	}
-
-	qry := cfg.Query()
-
-	// Force PostgreSQL session timezone to UTC.
-	qry.Set("timezone", "UTC")
-
-	// Force application name.
-	qry.Set("application_name", app)
-
-	// Set max open and idle connections
-	maxOpen, _ := strconv.Atoi(qry.Get("max_conns"))
-	if maxOpen == 0 {
-		maxOpen = 30
-	}
-	qry.Del("max_conns")
-
-	cfg.RawQuery = qry.Encode()
-	db, err := sql.Open("postgres", cfg.String())
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to connect to database")
-	}
-
-	// On sourcegraph.com we can sometimes startup faster than the
-	// cloud-sql-proxy.
-	if err := pingDbWithRetry(db, 10); err != nil {
-		return nil, errors.Wrap(err, "failed to ping database")
-	}
-
-	db.SetMaxOpenConns(maxOpen)
-	db.SetMaxIdleConns(maxOpen)
-	db.SetConnMaxLifetime(time.Minute)
-
-	return db, nil
-}
-
-func pingDbWithRetry(db *sql.DB, attempts int) error {
-	const maxWait = time.Second
-	wait := 50 * time.Millisecond
-	for i := 0; i < attempts-1; i++ {
-		if err := db.Ping(); err == nil {
-			return nil
-		}
-		time.Sleep(wait)
-		wait *= 2
-		if wait > maxWait {
-			wait = maxWait
-		}
-	}
-	return db.Ping()
-}
-
-// databases configures the migrations we want based on a database name. This
-// configuration includes the name of the migration version table as well as
-// the raw migration assets to run to migrate the target schema to a new version.
-var databases = map[string]struct {
-	MigrationsTable string
-	Resource        *bindata.AssetSource
-}{
-	"frontend": {
-		MigrationsTable: "schema_migrations",
-		Resource:        bindata.Resource(frontendMigrations.AssetNames(), frontendMigrations.Asset),
-	},
-	"codeintel": {
-		MigrationsTable: "codeintel_schema_migrations",
-		Resource:        bindata.Resource(codeintelMigrations.AssetNames(), codeintelMigrations.Asset),
-	},
-}
-
-// DatabaseNames returns the list of database names (configured via `dbutil.databases`)..
-var DatabaseNames = func() []string {
-	var names []string
-	for databaseName := range databases {
-		names = append(names, databaseName)
-	}
-
-	return names
-}()
-
-// MigrationTables returns the list of migration table names (configured via `dbutil.databases`).
-var MigrationTables = func() []string {
-	var migrationTables []string
-	for _, db := range databases {
-		migrationTables = append(migrationTables, db.MigrationsTable)
-	}
-
-	return migrationTables
-}()
-
-// NewMigrate returns a new configured migration object for the given database name. This database
-// name must be present in the `dbutil.databases` map. This migration can be subsequently run by
-// invoking `dbutil.DoMigrate`.
-func NewMigrate(db *sql.DB, databaseName string) (*migrate.Migrate, error) {
-	schemaData, ok := databases[databaseName]
-	if !ok {
-		return nil, fmt.Errorf("unknown database '%s'", databaseName)
-	}
-
-	driver, err := postgres.WithInstance(db, &postgres.Config{
-		MigrationsTable: schemaData.MigrationsTable,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	d, err := bindata.WithInstance(schemaData.Resource)
-	if err != nil {
-		return nil, err
-	}
-
-	m, err := migrate.NewWithInstance("go-bindata", d, "postgres", driver)
-	if err != nil {
-		return nil, err
-	}
-
-	// In case another process was faster and runs migrations, we will wait
-	// this long
-	m.LockTimeout = 5 * time.Minute
-	if os.Getenv("LOG_MIGRATE_TO_STDOUT") != "" {
-		m.Log = stdoutLogger{}
-	}
-
-	return m, nil
-}
-
-// DoMigrate runs all up migrations.
-func DoMigrate(m *migrate.Migrate) (err error) {
-	err = m.Up()
-	if err == nil || err == migrate.ErrNoChange {
-		return nil
-	}
-
-	if os.IsNotExist(err) {
-		// This should only happen if the DB is ahead of the migrations available
-		version, dirty, verr := m.Version()
-		if verr != nil {
-			return verr
-		}
-		if dirty { // this shouldn't happen, but checking anyways
-			return err
-		}
-		log15.Warn("WARNING: Detected an old version of Sourcegraph. The database has migrated to a newer version. If you have applied a rollback, this is expected and you can ignore this warning. If not, please contact support@sourcegraph.com for further assistance.", "db_version", version)
-		return nil
-	}
-	return err
 }
 
 type stdoutLogger struct{}
